Made verbose messages for stopping listener less confusing
Details:

* Made the code in run_exit_handler() (exit handler in the 'run' process)
  that sends the failure signal back to the 'start' process dependent on
  whether the start process still exists. This avoids confusing verbose
  messages in the normal shutdown of the 'run' process, at which point
  the 'start' process has long ceased to exist.

  This is just an optimization to remove confusing verbose messages, and
  not a fix specifically for Windows.

+++ pywbemtools/pywbemlistener/_cmd_listener.py
@@ -620,21 +620,29 @@ def run_exit_handler(start_pid, log_fp):
 
     In addition, it closes the log_fp log file.
     """
-    if _config.VERBOSE_PROCESSES_ENABLED:
-        print_out("Run process exit handler: Sending failure signal ({}) to "
-                  "start process {}".
-                  format(SIGNAL_RUN_STARTUP_FAILURE, start_pid))
     try:
-        os.kill(start_pid, SIGNAL_RUN_STARTUP_FAILURE)  # Sends the signal
-    except OSError:
-        # Note: ProcessLookupError is a subclass of OSError but was introduced
-        # only in Python 3.3.
+        start_p = psutil.Process(start_pid)
+    except psutil.NoSuchProcess:
+        start_p = None
 
-        # The original start parent no longer exists -> the earlier startup
-        # succeeded.
+    if start_p:
         if _config.VERBOSE_PROCESSES_ENABLED:
-            print_out("Run process exit handler: Start process {} does not "
-                      "exist anymore".format(start_pid))
+            print_out("Run process exit handler: Sending failure signal ({}) "
+                      "to start process {}".
+                      format(SIGNAL_RUN_STARTUP_FAILURE, start_pid))
+        try:
+            os.kill(start_pid, SIGNAL_RUN_STARTUP_FAILURE)  # Sends the signal
+        except OSError:
+            # Note: ProcessLookupError is a subclass of OSError but was
+            # introduced only in Python 3.3.
+
+            # The original start parent no longer exists.
+            # This can only happen if the process goes away in the short time
+            # window between checking for it at the begin of this function,
+            # and here.
+            if _config.VERBOSE_PROCESSES_ENABLED:
+                print_out("Run process exit handler: Start process {} does "
+                          "not exist anymore".format(start_pid))
 
     if log_fp:
         print_out("Closing 'run' output log file at {}".format(datetime.now()))
@@ -835,11 +843,13 @@ def run_term_signal_handler(sig, frame):
     SIGTERM signal, i.e. when the 'run' process gets terminated by
     some means.
 
-    Ths handler ensures that the main loop of the the 'run' command
+    This handler ensures that the main loop of the the 'run' command
     gets control and can gracefully stop the listener.
     """
     if _config.VERBOSE_PROCESSES_ENABLED:
         print_out("Run process: Received termination signal ({})".format(sig))
+
+    # This triggers the registered exit handler run_exit_handler()
     raise SystemExit(1)
 
 
@@ -1148,10 +1158,9 @@ def cmd_listener_start(context, name, options):
 
     prepare_startup_completion()
 
-    if six.PY2:
-        popen_kwargs = dict(shell=False)
-    else:
-        popen_kwargs = dict(shell=False, start_new_session=True)
+    popen_kwargs = dict(shell=False)
+    if six.PY3:
+        popen_kwargs['start_new_session'] = True
 
     if _config.VERBOSE_PROCESSES_ENABLED:
         print_out("Start process {}: Starting run process as: {}".
